+++ packages/core/src/redux/rootReducer.ts
@@ -1,4 +1,4 @@
-import { combineReducers } from './core'
+import { combineReducers } from './toolkit'
 import {
   componentReducer,
   sessionReducer,
